Improve Code

+++ DropFiles.h
@@ -30,7 +30,7 @@ private:
 
 	#pragma region Functions
 	static VOID CreateDropFilesFont();
-	static VOID drawDashedRectangle(HDC hdc, RECT &Rectangle, SIZE_T Width, COLORREF Color);
+	static VOID drawDashedRectangle(HDC hdc, RECT &Rectangle, UINT Width, COLORREF Color);
 	static VOID FillRectOpacity50(HDC hdc, RECT &Rectangle,COLORREF Color);
 	static VOID drawArrow(HDC hdc, INT COORD_X, INT COORD_Y, INT WIDTH = 20, INT HEIGHT = 20, COLORREF Color = BLACK_COLOR);
 	#pragma endregion
